pyproject.toml
[project]
name = "mcp-keycloak"
version = "0.1.0"
description = "MCP server for Keycloak administration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.3",
    "requests>=2.32.0",
]

[project.scripts]
mcp-keycloak = "mcp_keycloak.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.wheel.force-include]
"src" = "mcp_keycloak"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.12.7",
]
